# storage/engine: reduce allocations in mvccPutInternal