Move maybeGetValue to a top-level function to avoid allocations that
occur when capturing variables. No perf difference on the micro
benchmarks, but the allocations show up on heap profiles and add to GC
pressure.

name                 old time/op    new time/op    delta
MVCCPut10_RocksDB-8    3.98µs ± 1%    3.97µs ± 1%      ~      (p=0.530 n=8+10)

name                 old allocs/op  new allocs/op  delta
MVCCPut10_RocksDB-8      3.00 ± 0%     0.00 ±NaN%  -100.00%  (p=0.000 n=10+10)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/13407)
<!-- Reviewable:end -->